part2/requirements.txt
paho-mqtt==1.5.0
sslkeylog==0.2.0
